django_project/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
blis==0.7.9
boto3==1.28.17
botocore==1.31.17
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.1.0
cymem==2.0.7
Django==4.1.7
django-bootstrap-v5==1.0.11
django-extensions==3.2.1
django-redis==5.2.0
django-storages==1.13.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
environs==9.5.0
frozenlist==1.3.3
hiredis==2.2.2
idna==3.4
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
JPype1==1.4.1
ko-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/ko_core_news_sm-3.5.0/ko_core_news_sm-3.5.0-py3-none-any.whl#sha256=383ce09dc37ee707d8db0e9321e9b081749000ab4250369738f62cbb79265b63
langcodes==3.3.0
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.19.0
multidict==6.0.4
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.0
openai==0.27.8
packaging==23.0
pathy==0.10.2
preshed==3.0.8
psycopg2-binary==2.9.5
py-hanspell==1.1
py4j==0.10.9.2
pyasn1==0.4.8
pydantic==1.10.11
PyKomoran==0.1.6.post1
pyspellchecker==0.7.2
python-dateutil==2.8.2
python-dotenv==0.21.1
python-rake==1.5.0
rake-nltk==1.0.6
redis==4.5.1
regex==2023.6.3
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy-legacy==3.0.12
spacy-loggers==1.0.4
sqlparse==0.4.3
srsly==2.4.6
textblob==0.17.1
thinc==8.1.10
tqdm==4.65.0
typer==0.9.0
typing_extensions==4.7.1
urllib3==1.26.16
wasabi==1.1.2
yarl==1.9.2
